+++ thinklab/thinklab/models/multimodal/facebook/sam3/model.py
@@ -62,7 +62,7 @@ class CLIPAttention(nn.Module):
         attn_w = (q @ k.transpose(-2, -1)) * self.scale
         if attn_mask is not None:
             attn_w = attn_w + attn_mask
-        attn_w = F.softmax(attn_w, dim=-1)
+        attn_w = F.softmax(attn_w, dim=-1).to(v.dtype)
         out = (attn_w @ v).transpose(1, 2).reshape(B, N, -1)
         return self.out_proj(out)
